Add the dashboard and its documentation.

+++ purge.sh
@@ -1,7 +1,35 @@
-docker kill $(docker ps -q)
+# Purge all docker containers & database values. ONLY USE IN DEV!!!
 
-docker rm $(docker ps -a -q)
+set -e
 
-docker rmi $(docker images -q)
+echo "Purging the Wywy-Website repository will remove docker containers, docker images, and Wywy-Website databases."
 
-docker system prune -af
+read -p "Do you want to proceed? [y/n] " yn
+
+if [[ ! "$yn" =~ [Yy]$ ]]; then
+    echo "Aborting."
+    exit 0
+fi
+
+echo "Proceeding with purge in 3,"
+sleep 1
+echo "2,"
+sleep 1
+echo "1."
+sleep 1
+echo "Beginning purge."
+sleep 0.1
+
+for service_name in $(cat /etc/Wywy-Website-Control/services.txt | cut -d',' -f1); do
+    read -p "Purge service $service_name? [y/n] " purge
+    if [[ ! "$purge" =~ ^[Yy]$ ]]; then
+        continue
+    fi
+
+    "/etc/Wywy-Website-Control/scripts/purge/$service_name.sh"
+done
+
+read -p "Prune docker? [y/n] " prune
+if [[ "$prune" =~ ^[Yy]$ ]]; then
+    docker system prune -af
+fi
